requirements.txt
# Основные библиотеки для Telegram бота
python-telegram-bot[rate-limiter,job-queue]==21.9

# База данных
aiosqlite==0.20.0

# Конфигурация и переменные окружения
python-dotenv==1.0.1

# HTTP клиенты и API
aiohttp==3.10.5
requests==2.31.0

# SSH и удаленное управление серверами
paramiko==3.4.0

# QR коды для конфигураций
qrcode==7.4.2
pypng

# Обработка изображений
Pillow==10.4.0

# Веб-интерфейсы
Flask==3.0.3
fastapi==0.115.0
uvicorn==0.32.0

# Флаги стран
emoji-country-flag==1.3.2

# Дополнительные зависимости для стабильной работы
cryptography>=41.0.0
bcrypt>=4.0.0
pynacl>=1.5.0
